+++ .eslintignore
@@ -1,3 +1,4 @@
 wallaby.conf.js
 dist
 src/app.test.js
+src/app-setup.test.js
