+++ src/app-setup.js
@@ -1,0 +1,103 @@
+const {
+  API,
+  APP,
+  PORT,
+  SESSION_TABLE_NAME,
+  LOG_LEVEL,
+  SESSION_SECRET,
+  SESSION_TTL
+} = require("./lib/config");
+
+const commonExpress = require("@govuk-one-login/di-ipv-cri-common-express");
+
+const { setGTM } = commonExpress.lib.settings;
+const { setAPIConfig, setOAuthPaths } = require("./lib/settings");
+const { setup } = require("hmpo-app");
+const path = require("path");
+const AWS = require("aws-sdk");
+const helmetConfig = commonExpress.lib.helmet;
+const setHeaders = commonExpress.lib.headers;
+const session = require("express-session");
+const DynamoDBStore = require("connect-dynamodb")(session);
+
+const init = (app) => {
+  setAPIConfig({
+    app,
+    baseUrl: API.BASE_URL,
+    sessionPath: API.PATHS.SESSION,
+    authorizationPath: API.PATHS.AUTHORIZATION
+  });
+
+  setOAuthPaths({ app, entryPointPath: APP.PATHS.FRAUD });
+
+  setGTM({
+    app,
+    ga4ContainerId: APP.GTM.GA4_ID,
+    uaContainerId: APP.GTM.UA_ID,
+    analyticsCookieDomain: APP.GTM.ANALYTICS_COOKIE_DOMAIN,
+    ga4Disabled: APP.GTM.GA4_DISABLED,
+    uaDisabled: APP.GTM.UA_DISABLED
+  });
+};
+
+const create = () => {
+  const loggerConfig = {
+    consoleLevel: LOG_LEVEL,
+    console: true,
+    consoleJSON: true,
+    app: false
+  };
+
+  AWS.config.update({
+    region: "eu-west-2"
+  });
+  const dynamodb = new AWS.DynamoDB();
+
+  const dynamoDBSessionStore = new DynamoDBStore({
+    client: dynamodb,
+    table: SESSION_TABLE_NAME
+  });
+
+  const sessionConfig = {
+    cookieName: "service_session",
+    secret: SESSION_SECRET,
+    cookieOptions: { maxAge: SESSION_TTL },
+    ...(SESSION_TABLE_NAME && { sessionStore: dynamoDBSessionStore })
+  };
+
+  const { app, router } = setup({
+    config: { APP_ROOT: __dirname },
+    port: PORT,
+    logs: loggerConfig,
+    session: sessionConfig,
+    helmet: helmetConfig,
+    redis: SESSION_TABLE_NAME ? false : commonExpress.lib.redis(),
+    urls: {
+      public: "/public",
+      publicImages: "/public/images"
+    },
+    publicDirs: ["../dist/public"],
+    publicImagesDirs: ["../dist/public/images"],
+    translation: {
+      allowedLangs: ["en", "cy"],
+      fallbackLang: ["en"],
+      cookie: { name: "lng" }
+    },
+    views: [
+      path.resolve(
+        path.dirname(
+          require.resolve("@govuk-one-login/di-ipv-cri-common-express")
+        ),
+        "components"
+      ),
+      "views"
+    ],
+    middlewareSetupFn: (app) => {
+      app.use(setHeaders);
+    },
+    dev: true
+  });
+  return { app, router };
+};
+
+module.exports = { init, create };
